Moved jtag.h and jtag_device.h to a separate folder
