+++ device/api/umd/device/tt_device/tt_device.h
@@ -14,7 +14,7 @@
 #include "umd/device/arc/arc_telemetry_reader.h"
 #include "umd/device/architecture_implementation.h"
 #include "umd/device/chip_helpers/tlb_manager.h"
-#include "umd/device/jtag_device.h"
+#include "umd/device/jtag/jtag_device.h"
 #include "umd/device/pci_device.hpp"
 #include "umd/device/types/cluster_descriptor_types.h"
 #include "umd/device/utils/lock_manager.h"
