+++ device/jtag/jtag.cpp
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
 //
 // SPDX-License-Identifier: Apache-2.0
-#include "umd/device/jtag.h"
+#include "umd/device/jtag/jtag.h"
 
 #include <dlfcn.h>
 #include <stdint.h>
