+++ device/tt_device/tt_device.cpp
@@ -11,7 +11,7 @@
 #include "assert.hpp"
 #include "umd/device/arc/arc_messenger.h"
 #include "umd/device/driver_atomics.h"
-#include "umd/device/jtag_device.h"
+#include "umd/device/jtag/jtag_device.h"
 #include "umd/device/pci_device.hpp"
 #include "umd/device/tt_device/blackhole_tt_device.h"
 #include "umd/device/tt_device/wormhole_tt_device.h"
