+++ device/utils/lock_manager.cpp
@@ -22,18 +22,12 @@ const std::unordered_map<MutexType, std::string> LockManager::MutexTypeToString 
 };
 
 const std::unordered_map<IODeviceType, std::string> LockManager::DeviceTypeToString = {
-    {IODeviceType::PCIe, "TT_PCIe"},
-    {IODeviceType::JTAG, "TT_JTAG"},
+    {IODeviceType::PCIe, "PCIe"},
+    {IODeviceType::JTAG, "JTAG"},
 };
 
 void LockManager::initialize_mutex(MutexType mutex_type) {
     initialize_mutex_internal(MutexTypeToString.at(mutex_type));
-}
-
-void LockManager::clear_mutex(MutexType mutex_type) { clear_mutex_internal(MutexTypeToString.at(mutex_type)); }
-
-std::unique_lock<RobustMutex> LockManager::acquire_mutex(MutexType mutex_type) {
-    return acquire_mutex_internal(MutexTypeToString.at(mutex_type));
 }
 
 void LockManager::initialize_mutex(MutexType mutex_type, int device_id, IODeviceType device_type) {
@@ -42,10 +36,16 @@ void LockManager::initialize_mutex(MutexType mutex_type, int device_id, IODevice
     initialize_mutex_internal(mutex_name);
 }
 
+void LockManager::clear_mutex(MutexType mutex_type) { clear_mutex_internal(MutexTypeToString.at(mutex_type)); }
+
 void LockManager::clear_mutex(MutexType mutex_type, int device_id, IODeviceType device_type) {
     std::string mutex_name =
         MutexTypeToString.at(mutex_type) + "_" + std::to_string(device_id) + "_" + DeviceTypeToString.at(device_type);
     clear_mutex_internal(mutex_name);
+}
+
+std::unique_lock<RobustMutex> LockManager::acquire_mutex(MutexType mutex_type) {
+    return acquire_mutex_internal(MutexTypeToString.at(mutex_type));
 }
 
 std::unique_lock<RobustMutex> LockManager::acquire_mutex(
